Improve integration of attribute option images
Using 'Mage_Eav' might not be the best approach to provide additional 
data. In this case it is needed for product attributes and thus primary 
for 'Mage_Catalog'. In fact the catalog attribute model extends 'Mage_Eav' 
already, by using an additional data table. This extends the 'catalog 
attribute option model' in a similiar way.

Thus there is no need anymore to use uncached resources through data 
helper in the frontend code. Instead, you can now go with the most 
common way, by using the default attribute source model (getAllOptions).

Further, the deprecated resource namespace 'Mysql4' is not longer used, 
which might be a problem for older versions of Magento. Data of existing 
installations would be transferred to the new table. And the field name 
for the thumbnail image has changed from 'thumb' to 'thumbnail'.